test: move dag-scope tests to their own file
